Rename notifier module to slack to better reflect upcoming changes
