+++ src/adapters/slack.ts
